fix(persistence): Data extraction efficiency
- Use smaller data types for columns
- Adds index to improve searches between dates
- Adds index to improve sort descending by priority
- Restrict the projection avoiding * in SELECT

+++ src/main/java/com/rodriguezblanco/priceservice/prices/infrastructure/persistence/entity/PriceIdentity.java
@@ -5,13 +5,13 @@ import java.util.Objects;
 
 public class PriceIdentity implements Serializable {
     private Long productId;
-    private Integer brandId;
-    private Integer priceList;
+    private Short brandId;
+    private Short priceList;
 
     public PriceIdentity() {
     }
 
-    public PriceIdentity(Long productId, Integer brandId, Integer priceList) {
+    public PriceIdentity(Long productId, Short brandId, Short priceList) {
         this.productId = productId;
         this.brandId = brandId;
         this.priceList = priceList;
